Move isDefined to be a method of MacroExpander
