+++ src/MacroExpander.js
@@ -4,6 +4,8 @@
  * until only non-macro tokens remain.
  */
 
+import functions from "./functions";
+import symbols from "./symbols";
 import Lexer from "./Lexer";
 import {Token} from "./Token";
 import type {Mode} from "./types";
@@ -14,6 +16,16 @@ import builtinMacros from "./macros";
 import type {MacroContextInterface, MacroDefinition, MacroExpansion}
     from "./macros";
 import type Settings from "./Settings";
+
+// List of commands that act like macros but aren't defined as a macro,
+// function, or symbol.  Used in `isDefined`.
+export const implicitCommands = {
+    "\\relax": true,     // MacroExpander.js
+    "^": true,           // Parser.js
+    "_": true,           // Parser.js
+    "\\limits": true,    // Parser.js
+    "\\nolimits": true,  // Parser.js
+};
 
 export default class MacroExpander implements MacroContextInterface {
     maxExpand: number;
@@ -281,6 +293,20 @@ export default class MacroExpander implements MacroContextInterface {
         }
 
         return expansion;
+    }
+
+    /**
+     * Determine whether a command is currently "defined" (has some
+     * functionality), meaning that it's a macro (in the current group),
+     * a function, a symbol, or one of the special commands listed in
+     * `implicitCommands`.
+     */
+    isDefined(name: string): boolean {
+        return this.macros.get(name) != null ||
+            functions.hasOwnProperty(name) ||
+            symbols.math.hasOwnProperty(name) ||
+            symbols.text.hasOwnProperty(name) ||
+            implicitCommands.hasOwnProperty(name);
     }
 }
